+++ packages/dara-core/changelog.md
@@ -4,7 +4,7 @@ title: Changelog
 
 ## NEXT
 
--  Added a new `Fallback.Custom` option which allows you to specify a custom component to render when while suspended
+-  Added a new `Fallback.Custom` option which allows you to specify a custom component to render when while suspended. The `fallback=` props for components now also supports passing any non-fallback component directly which implicitly wraps it in `Fallback.Custom`
 
 ## 1.18.3